requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.8
influxalchemy==0.3.0
influxdb==5.3.2
influxdb-client==1.45.0
msgpack==1.0.8
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
PyQt5-Qt5==5.15.14
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reactivex==4.0.4
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
questdb~=2.0.3